backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==1.2.0
pandas==2.0.3
python-dotenv==1.0.0
tensorflow==2.13.0
joblib==1.3.1
gunicorn==21.2.0